backend/requirements.txt
Flask
Flask-CORS
pymongo
psycopg2
python-dotenv
pandas
torch
joblib
numpy
scikit-learn